% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refint.R
\name{predict.refint}
\alias{predict.refint}
\title{Predict Method for Reference Intervals}
\usage{
\method{predict}{refint}(object, newdata, ...)
}
\arguments{
\item{object}{\code{refint} object obtained from \link[ReForm]{refint}}

\item{newdata}{input variables used to get reference intervals}

\item{...}{additional arguments passed to \link[stats]{predict}}
}
\value{
a data frame containing lower and upper bounds of the reference
interval, the test observation, and logicals for whether it is above or below
}
\description{
Apply fitted reference interval from \link[ReForm]{refint} to new data
}
