% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reform.R
\name{reform}
\alias{reform}
\title{Reference interval calibration via conFormal prediction (ReForm)}
\usage{
reform(object, caldata, ...)
}
\arguments{
\item{object}{\code{refint} object obtained from \link[ReForm]{refint}}

\item{caldata}{calibration data used to fit ReForm}

\item{...}{additional arguments passed internally to \link[stats]{predict}}
}
\value{
\code{reform} returns a list containing all elements of \code{object} (see \link[ReForm]{refint}) and:
\item{cali}{lower and upper calibration constants used to adjust the interval}
\item{cali.df}{\code{caldata} stored in the output}
}
\description{
\code{reform} takes a fitted reference interval and calibrates to a new sample
using split conformalized quantile regression (Romano, et al., 2019).
\code{reform} requires a calibration set \code{caldata} from the new sample. For more
details on calibration set size and ReForm properties, refer to our preprint
Chen et al., 2025.
}
\examples{
# example for lm (not recommended)
ri <- refint(
  lm(Sepal.Length ~ Petal.Length + Species,
    data = iris[-(1:45),]), pct = 90
)
reformed_ri <- reform(ri, caldata = iris[1:40,])
predict(reformed_ri, newdata = iris[41:45,])

plot(reformed_ri)

# example for quantreg::rq
if (require("quantreg")) {
  ri <- refint(
    rq(Sepal.Length ~ Petal.Length,
      data = iris[-(1:45),], tau = 0.05),
    rq(Sepal.Length ~ Petal.Length,
      data = iris[-(1:45),], tau = 0.95)
  )
  reformed_ri <- reform(ri, caldata = iris[1:40,])
  predict(reformed_ri, newdata = iris[41:45,])

  plot(reformed_ri)
}

}
\references{
Romano, Y., Patterson, E., & Candes, E. (2019). Conformalized quantile regression. Advances in neural information processing systems, 32.
}
\seealso{
\link[ReForm]{plot.reformint} for assessing fit of reference interval in
calibration set via diagnostic plots

\link[ReForm]{predict.reformint} for applying a ReFormed reference
interval to new observations
}
