% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reform.R
\name{plot.reformint}
\alias{plot.reformint}
\title{Diagnostic Plots for ReForm}
\usage{
\method{plot}{reformint}(x, var = NULL, ...)
}
\arguments{
\item{x}{\code{reformint} object obtained from \link[ReForm]{reform}}

\item{var}{string indicating to plot against variable to plot against (should
be in \code{x$terms}). If none supplied, defaults to the first covariate (second
value) in \code{x$terms}.}

\item{...}{additional arguments passed to \link[base]{plot}}
}
\description{
Plot upper and lower residuals from the calibration set in
\link[ReForm]{reform} against a chosen variable \code{var}. Ideally, both plots
should display a roughly flat relationship.
}
\examples{
# example for lm (not recommended)
ri <- refint(
  lm(Sepal.Length ~ Petal.Length + Species,
    data = iris[-(1:40),])
)
reformed_ri <- reform(ri, caldata = iris[1:30,])
plot(reformed_ri)
}
