% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reform.R
\name{predict.reformint}
\alias{predict.reformint}
\title{Predict Method for ReFormed Reference Intervals}
\usage{
\method{predict}{reformint}(object, newdata, ...)
}
\arguments{
\item{object}{\code{reformint} object obtained from \link[ReForm]{reform}}

\item{newdata}{input variables used to get reference intervals}

\item{...}{additional arguments passed to \link[stats]{predict}}
}
\value{
a data frame containing lower and upper bounds of the reference
interval, the test observation, and logicals for whether it is above or below
}
\description{
Apply ReFormed reference interval from \link[ReForm]{reform} to new data
}
\examples{
# example for lm (not recommended)
ri <- refint(
  lm(Sepal.Length ~ Petal.Length + Species,
    data = iris[-(1:45),]), pct = 90
)
reformed_ri <- reform(ri, caldata = iris[1:40,])
predict(reformed_ri, newdata = iris[41:45,])

plot(reformed_ri)

# example for quantreg::rq
if (require("quantreg")) {
  ri <- refint(
    rq(Sepal.Length ~ Petal.Length,
      data = iris[-(1:45),], tau = 0.05),
    rq(Sepal.Length ~ Petal.Length,
      data = iris[-(1:45),], tau = 0.95)
  )
  reformed_ri <- reform(ri, caldata = iris[1:40,])
  predict(reformed_ri, newdata = iris[41:45,])

  plot(reformed_ri)
}
}
