% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reform.R
\name{reform}
\alias{reform}
\title{Reference interval calibration via conFormal prediction (ReForm)}
\usage{
reform(x, newdata, ...)
}
\arguments{
\item{x}{\code{refint} object, typically result of \link[ReForm]{refint}}

\item{newdata}{calibration data used to fit ReForm}

\item{...}{other parameters passed internally to \code{predict}}
}
\description{
Reference interval calibration via conFormal prediction (ReForm)
}
\examples{
# example for gamlss2
library(gamlss2)
fit <- gamlss2(Sepal.Length ~ Petal.Width + Species, data = iris[1:110,])
ri <- refint(fit, 95)
reformint <- reform(ri, newdata = iris[111:130,])
}
\seealso{
\link[ReForm]{plot.reformint} for assessing fit of reference interval in
calibration set via diagnostic plots

\link[ComBatFamily]{predict.reformint} for applying a ReFormed reference
interval to new observations
}
