% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reform.R
\name{reform}
\alias{reform}
\title{Reference interval calibration via conFormal prediction (ReForm)}
\usage{
reform(x, caldata, ...)
}
\arguments{
\item{x}{\code{refint} object obtained from \link[ReForm]{refint}}

\item{caldata}{calibration data used to fit ReForm}

\item{...}{additional arguments passed internally to \link[stats]{predict}}
}
\description{
\code{reform} takes a fitted reference interval and calibrates to a new sample
using split conformalized quantile regression (Romano, et al., 2019).
\code{reform} requires a calibration set from the new sample (\code{newdata}). For
more details on calibration set size and ReForm properties, refer to our
preprint Chen et al., 2025.
}
\examples{
# example for lm (not recommended)
ri <- refint(
  lm(Sepal.Length ~ Petal.Width + Species,
    data = iris[-(1:40),])
)
reformed_ri <- reform(ri, caldata = iris[1:30,])
predict(reformed_ri, newdata = iris[31:40,])

# example for quantreg::rq
if (require("quantreg")) {
  ri <- refint(
    rq(Sepal.Length ~ Petal.Length,
      data = iris[1:50,], tau = 0.025),
    rq(Sepal.Length ~ Petal.Length,
      data = iris[1:50,], tau = 0.975)
  )
  reformed_ri <- reform(ri, caldata = iris[1:30,])
  predict(reformed_ri, newdata = iris[31:40,])
}

}
\references{
Romano, Y., Patterson, E., & Candes, E. (2019). Conformalized quantile regression. Advances in neural information processing systems, 32.
}
\seealso{
\link[ReForm]{plot.reformint} for assessing fit of reference interval in
calibration set via diagnostic plots

\link[ReForm]{predict.reformint} for applying a ReFormed reference
interval to new observations
}
